requirements.txt
flask
requests
python-dotenv
celery
redis
sentry-sdk
gunicorn
